+++ app/Http/Controllers/AgentController.php
@@ -9,56 +9,72 @@ use Illuminate\View\View;
 
 class AgentController extends Controller
 {
+    /**
+     * Display a listing of agents.
+     */
     public function index(): View
     {
-        $data['agents'] = Agent::query()->latest()->paginate(10);
+        $this->authorize('view::agent');
+        $data['agents'] = Agent::query()->latest()->get();
 
         return $this->extendedView('agents.index', $data, 'agents');
     }
 
+    /**
+     * Store a newly created agent in storage.
+     */
     public function store(StoreAgentRequest $request): RedirectResponse
     {
+        $this->authorize('create::agent');
         $validated = $request->validated();
 
-        Agent::create($validated);
+        Agent::query()->create($validated);
 
         return redirect()
             ->route('agents.index')
             ->with('success', 'Agent created successfully.');
     }
 
-    public function create(): View
+    public function create(): RedirectResponse
     {
-        return view('agents.create');
+        return back()->with('flush_error', 'Not Found');
     }
 
-    public function show(Agent $agent): View
+    public function show(Agent $agent): RedirectResponse
     {
-        return view('agents.show', compact('agent'));
+        return back()->with('flush_error', 'Not Found');
     }
 
-    public function edit(Agent $agent): View
+    public function edit(Agent $agent): RedirectResponse
     {
-        return view('agents.edit', compact('agent'));
+        return back()->with('flush_error', 'Not Found');
     }
 
+    /**
+     * Update the specified agent in storage.
+     */
     public function update(StoreAgentRequest $request, Agent $agent): RedirectResponse
     {
+        $this->authorize('edit::agent');
         $validated = $request->validated();
 
         $agent->update($validated);
 
         return redirect()
             ->route('agents.index')
-            ->with('success', 'Agent updated successfully.');
+            ->with('flush_success', 'Agent updated successfully.');
     }
 
+    /**
+     * Remove the specified agent from storage.
+     */
     public function destroy(Agent $agent): RedirectResponse
     {
+        $this->authorize('delete::agent');
         $agent->delete();
 
         return redirect()
             ->route('agents.index')
-            ->with('success', 'Agent deleted successfully.');
+            ->with('flush_success', 'Agent deleted successfully.');
     }
 }
