+++ app/Http/Controllers/LodgeController.php
@@ -9,15 +9,23 @@ use Illuminate\View\View;
 
 class LodgeController extends Controller
 {
+    /**
+     * Display a listing of lodges.
+     */
     public function index(): View
     {
+        $this->authorize('view::lodge');
         $data['lodges'] = Lodge::query()->latest()->paginate(10);
 
         return $this->extendedView('lodges.index', $data, 'lodges');
     }
 
+    /**
+     * Store a newly created lodge in storage.
+     */
     public function store(StoreLodgeRequest $request): RedirectResponse
     {
+        $this->authorize('create::lodge');
         $validated = $request->validated();
 
         Lodge::create($validated);
@@ -42,8 +50,12 @@ class LodgeController extends Controller
         return view('lodges.edit', compact('lodge'));
     }
 
+    /**
+     * Update the specified lodge in storage.
+     */
     public function update(StoreLodgeRequest $request, Lodge $lodge): RedirectResponse
     {
+        $this->authorize('edit::lodge');
         $validated = $request->validated();
 
         $lodge->update($validated);
@@ -53,8 +65,12 @@ class LodgeController extends Controller
             ->with('success', 'Lodge updated successfully.');
     }
 
+    /**
+     * Remove the specified lodge from storage.
+     */
     public function destroy(Lodge $lodge): RedirectResponse
     {
+        $this->authorize('delete::lodge');
         $lodge->delete();
 
         return redirect()
