+++ app/Http/Controllers/ServiceProviderController.php
@@ -7,8 +7,10 @@ use App\Http\Requests\StoreServiceProviderRequest;
 use App\Models\BankAccount;
 use App\Models\Country;
 use App\Models\ServiceProvider;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
+use Illuminate\View\View;
 use Throwable;
 
 class ServiceProviderController extends Controller
@@ -16,8 +18,9 @@ class ServiceProviderController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request)
+    public function index(Request $request): View
     {
+        $this->authorize('view::service_provider');
         $data['banks'] = BankName::cases();
         $data['countries'] = Country::all();
         $data['serviceproviders'] = ServiceProvider::all();
@@ -28,8 +31,9 @@ class ServiceProviderController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreServiceProviderRequest $request)
+    public function store(StoreServiceProviderRequest $request): RedirectResponse
     {
+        $this->authorize('create::service_provider');
         try {
             $validated = $request->validated();
             $validated['voucher_copies'] = 3;
@@ -77,8 +81,9 @@ class ServiceProviderController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(StoreServiceProviderRequest $request, $id)
+    public function update(StoreServiceProviderRequest $request, $id): RedirectResponse
     {
+        $this->authorize('edit::service_provider');
         try {
             $validated = $request->validated();
             $provider = ServiceProvider::findOrFail($id);
@@ -119,8 +124,9 @@ class ServiceProviderController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy($id): RedirectResponse
     {
+        $this->authorize('delete::service_provider');
         try {
             $provider = ServiceProvider::findOrFail($id);
             $provider->delete();
